# internal/transport/rest/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/prices/database-nosql": {
            "get": {
                "description": "Returns normalized NoSQL database pricing across cloud providers (AWS DynamoDB, Azure Cosmos DB, GCP Firestore) for requested specs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prices"
                ],
                "summary": "Get NoSQL database pricing comparison",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requested NoSQL data model (e.g. document, key_value, wide_column, graph)",
                        "name": "data_model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Requested pricing mode (provisioned, on_demand, serverless; default: provisioned)",
                        "name": "pricing_mode",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Requested reads per second (e.g. 100)",
                        "name": "read_units",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Requested writes per second (e.g. 20)",
                        "name": "write_units",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Requested storage in GB (e.g. 50)",
                        "name": "storage_gb",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Storage class preference (e.g. standard, infrequent_access, analytical)",
                        "name": "storage_class",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "High Availability / Multi-Region replication (default: false)",
                        "name": "multi_region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Canonical region group (default: us-east)",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Target currency code (default: USD)",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.DatabaseNoSQLComparisonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    }
                }
            }
        },
        "/api/v1/prices/kubernetes": {
            "get": {
                "description": "Returns normalized Kubernetes control-plane pricing across cloud providers (AWS EKS, Azure AKS, GCP GKE) for requested tier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prices"
                ],
                "summary": "Get Kubernetes control-plane pricing comparison",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Requested Kubernetes cluster tier (free, standard, extended_support; default: standard)",
                        "name": "tier",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "GCP-specific cluster topology for credit application (zonal, regional, autopilot; default: zonal)",
                        "name": "cluster_topology",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Canonical region group (default: us-east)",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Target currency code (default: USD)",
                        "name": "currency",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.KubernetesComparisonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.RFC7807Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DatabaseNoSQLAttributes": {
            "type": "object",
            "properties": {
                "component_type": {
                    "description": "\"throughput\", \"storage\", \"request_operations\"",
                    "type": "string"
                },
                "data_model": {
                    "description": "\"document\", \"key_value\", \"wide_column\", \"graph\", \"multi_model\"",
                    "type": "string"
                },
                "multi_region": {
                    "description": "Multi-region replication / global distribution",
                    "type": "boolean"
                },
                "pricing_mode": {
                    "description": "\"provisioned\", \"on_demand\", \"serverless\"",
                    "type": "string"
                },
                "read_units": {
                    "description": "Provisioned reads/sec (RCU/RU/s) or request rate",
                    "type": "number"
                },
                "replication_zones": {
                    "description": "Number of replica regions (default: 1)",
                    "type": "integer"
                },
                "storage_class": {
                    "description": "\"standard\", \"infrequent_access\", \"analytical\"",
                    "type": "string"
                },
                "storage_gb": {
                    "description": "Storage capacity in GiB",
                    "type": "number"
                },
                "write_units": {
                    "description": "Provisioned writes/sec (WCU/RU/s) or request rate",
                    "type": "number"
                }
            }
        },
...
        "domain.KubernetesAttributes": {
            "type": "object",
            "properties": {
                "cluster_topology": {
                    "description": "\"zonal\", \"regional\", \"autopilot\" — GCP-specific",
                    "type": "string"
                },
                "tier": {
                    "description": "\"free\", \"standard\", \"extended_support\"",
                    "type": "string"
                }
            }
        },
...
        "middleware.RFC7807Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "rest.DatabaseNoSQLComparisonMeta": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "generated_at": {
                    "type": "string"
                },
                "query": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "rest.DatabaseNoSQLComparisonResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/rest.DatabaseNoSQLComparisonMeta"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.DatabaseNoSQLResultEntry"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ProviderWarning"
                    }
                }
            }
        },
        "rest.DatabaseNoSQLResultEntry": {
            "type": "object",
            "properties": {
                "fetched_at": {
                    "type": "string"
                },
                "match_delta_pct": {
                    "type": "number"
                },
                "match_quality": {
                    "type": "string"
                },
                "matched_spec": {
                    "$ref": "#/definitions/domain.DatabaseNoSQLAttributes"
                },
                "missing_attributes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "normalized_hourly_usd": {
                    "type": "number"
                },
                "price": {
                    "$ref": "#/definitions/rest.PriceDetail"
                },
                "provider": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "string"
                },
                "stale": {
                    "type": "boolean"
                }
            }
        },
...
        "rest.KubernetesComparisonMeta": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "generated_at": {
                    "type": "string"
                },
                "query": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "rest.KubernetesComparisonResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/rest.KubernetesComparisonMeta"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.KubernetesResultEntry"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.ProviderWarning"
                    }
                }
            }
        },
        "rest.KubernetesResultEntry": {
            "type": "object",
            "properties": {
                "fetched_at": {
                    "type": "string"
                },
                "match_delta_pct": {
                    "type": "number"
                },
                "match_quality": {
                    "type": "string"
                },
                "matched_spec": {
                    "$ref": "#/definitions/domain.KubernetesAttributes"
                },
                "missing_attributes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "normalized_hourly_usd": {
                    "type": "number"
                },
                "price": {
                    "$ref": "#/definitions/rest.PriceDetail"
                },
                "provider": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "string"
                },
                "stale": {
                    "type": "boolean"
                }
            }
        },
...
        "rest.PriceDetail": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "rest.ProviderWarning": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
    }
}